Add configure() function to allow customizing labels
A configure() function is added to the main entry point, allowing
global configuration of the module.
The only configuration supported is labels to use for the two compared
versions, which defaults to 'expected' and 'actual'. The labels are
given as an array of the two labels.

The configure function returns the entry point to allow chaining:

const printDiff = require('print-diff').configure({labels: ['before', 'after']});
printDiff('Hello\nLinus', 'Hello\nworld')